05-nest-js/src/domain/forum/application/use-cases/delete-answer/index.ts
import { Either, left, right } from '@/core/error/either'
import { AnswersRepository } from '../../repositories/answers'
import { NotAllowedError } from '../_errors/not-allowed-error'
import { ResourceNotFoundError } from '../_errors/resource-not-found'

export type TDeleteAnswerUseCaseRequest = {
  answerId: string
  authorId: string
}

export type TDeleteAnswerUseCaseResponse = Either<
  ResourceNotFoundError | NotAllowedError,
  null
>

export class DeleteAnswerUseCase {
  constructor(private answersRepository: AnswersRepository) {}

  async execute({
    answerId,
    authorId,
  }: TDeleteAnswerUseCaseRequest): Promise<TDeleteAnswerUseCaseResponse> {
    const answer = await this.answersRepository.findById(answerId)

    if (!answer) {
      return left(new ResourceNotFoundError())
    }

    if (authorId !== answer.authorId.toString()) {
      return left(new NotAllowedError())
    }

    await this.answersRepository.delete(answer)

    return right(null)
  }
}
